fix(*): few updates and fixes
